requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
deprecation==2.1.0
docker==6.0.1
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.1
requests==2.28.2
SQLAlchemy==2.0.4
testcontainers==3.7.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
wrapt==1.14.1